Leave card: > 1. Merge pull request from Julia > 2. Add method deleteUserFromCard to DefaultCardUserService > 3. Add test deleteUserWithError to DefaultCardUserServiceMockITest > 4. Move methode leaveCard to CardUserController
