+++ src/main/java/com/greetingcard/service/CardUserService.java
@@ -11,4 +11,6 @@ public interface CardUserService {
     List<UserInfo> getUsersByCardId(long cardId, User userLoggedIn);
 
     void deleteUsers(long cardId, List<UserInfo> listUserIds, User userLoggedIn);
+
+    void deleteUserFromCard(long cardId, long userId);
 }
