--- src/test/resources/datasets/cardsUsersAfterDeleteByCardId.xml
@@ -1,22 +1,0 @@
-<dataset>
-    <USERS_CARDS users_cards_id ="2"
-                 user_id = "2"
-                 card_id = "1"
-                 role_id = "2"/>
-    <USERS_CARDS users_cards_id ="3"
-                 user_id = "2"
-                 card_id = "2"
-                 role_id = "1"/>
-    <USERS_CARDS users_cards_id ="4"
-                 user_id = "1"
-                 card_id = "2"
-                 role_id = "2"/>
-    <USERS_CARDS users_cards_id ="5"
-                 user_id = "2"
-                 card_id = "3"
-                 role_id = "1"/>
-    <USERS_CARDS users_cards_id ="6"
-                 user_id = "1"
-                 card_id = "3"
-                 role_id = "2"/>
-</dataset>
