--- src/test/resources/dev.properties
@@ -1,3 +1,0 @@
-jdbc.url=jdbc:postgresql://localhost:5432/greeting_card
-jdbc.user=greeting_user
-jdbc.password=greeting
